Use Long as migration version

+++ src/main/kotlin/app/softwork/sqldelightwriter/SqWriter.kt
@@ -24,7 +24,7 @@ public class SqWriter(
 
     @SqDsl
     public fun migrationFile(
-        version: Int,
+        version: Long,
         packageName: String = this.packageName,
         migration: SqMigrationFile.() -> Unit
     ) {
